implemented detection of weathercard currently displayed
Because we want to be able to update pagination (highlighting a dot on a list of dots that refers to the number of weather cards open), we should create a state that's an array of booleans.

The array size should be the size of number of weather cards we have at open, and each value of the array should represent whether or not a certain card is displayed.

e.g. in a set of 3 items, if item at index 0 is shown, the corresponding truth array would be

[true, false, false]

this way, we can use this array to hydrate a pagination pattern that will clearly represent which card is being viewed.

Using react-intersection-observer, we are able to detect when elements are in viewport.

in WeatherCard.js (line 18) we create this observability object which takes in a ref (ref:cardRef) and returns a boolean to show whether or not that targeted ref is within view.

on the right-hand-side, we set it to useInView with a threshold of 0.5, or 50%. This number indicates how much of the object should be in the viewport before the effect triggers and sets the inView (cardIsVisible) to true.

Using this observability object, we can run an effect whenever cardIsVisible changes, so that we can update the original corresponding truth array:

1) pass down a setter function and WeatherCard index from App.js as a prop
2) call that function in the useEffect hook with the newly determined cardIsVisible state and the index of the card it was coming from
3) ensure that the dependency (cardIsVisible) is set so that the useEffect hook fires every time cardIsVisible changes because of viewport detection
4) in app.js, create a new array of truthiness values with an updated value on the index that was passed, otherwise push the same elements in the object