successfully populated a preview (CitiesOverviewTiles) with state set by API fetches in WeatherCard.js
We want to use the available data from WeatherData.js to populate a bunch of previews inside CitiesOverview.js for when the user hits the list button.

In App.js, create state empty state for citiesOverviewData with the length of the citiesLatLng array.

Pass the setter function into each WeatherCard component. Using the index of the item, populate a new array at that index with the necessary data (current temp, daily high/low, comment). Use the setter function and place the array in there.