Successfully added a new city to the Overview view after user queries a city.
set upperboundary to citiesLatLngLength, because prevCitiesOverviewData.length does not contain the updated length of the array that holds the cities' latlng objects.
citiesLatLngLength (citiesLatLng.length) is state that is dynamically updated based on the user completing the form in CitiesOverviewForm.jsx.
Querying a city name in text returns the latlng coordinates as an object, adds it to the citiesLatLng array, which then will update the citiesLatLng.length property.
after this, we want to update the citiesOverviewData to include preview information that helps hydrate the "overview" view.