feat - Successfully add refs to each WeatherCard and scrolls it into view when it's clicked
By creating a ref object and setting its value to [], we can store multiple elements that need to be tracked with a ref.

In this case, setting weatherCardRef.current = [], we can then set the ref to the element being mapped. (App.jsx line 132)

Splicing allows us to effectively insert the element that's being tracked (a rendered WeatherCard with data coming from data[index]).

This can be seen as an attachment of a ref to the rendered instance of WeatherCard.

We need the splice method because simply using push can result in nulls from unfetched data and therefore dehydrated && null components.

READ MORE: https://mattclaffey.medium.com/adding-react-refs-to-an-array-of-items-96e9a12ab40c